+++ src/RegistryAware.php
@@ -41,76 +41,23 @@
  * @license http://www.opensource.org/licenses/bsd-license.php BSD License
  */
 
-namespace SimplePie\Tests\Unit\Parse;
+namespace SimplePie;
 
-use PHPUnit\Framework\TestCase;
-use SimplePie\Parse\Date;
-
-class DateTest extends TestCase
+/**
+ * Handles the injection of Registry into other class
+ *
+ * {@see \SimplePie\SimplePie::get_registry()}
+ *
+ * @package SimplePie
+ */
+interface RegistryAware
 {
-    public function testNamespacedClassExists()
-    {
-        $this->assertTrue(class_exists('SimplePie\Parse\Date'));
-    }
-
-    public function testClassExists()
-    {
-        $this->assertTrue(class_exists('SimplePie_Parse_Date'));
-    }
-
-    public function w3cDtfDatesProvider()
-    {
-        // The examples enclosed within come from the W3C Date and Time Formats note
-        // https://www.w3.org/TR/NOTE-datetime
-        yield 'Year' => [
-            '1997',
-            852076800,
-        ];
-        yield 'Year and month' => [
-            '1997-07',
-            867715200,
-        ];
-        yield 'Complete date' => [
-            '1997-07-16',
-            869011200,
-        ];
-        yield 'Complete date plus hours and minutes' => [
-            '1997-07-16T19:20+01:00',
-            869077200,
-        ];
-        yield 'Complete date plus hours, minutes and seconds' => [
-            '1997-07-16T19:20:30+01:00',
-            869077230,
-        ];
-        yield 'Complete date plus hours, minutes, seconds and a decimal fraction of a second' => [
-            '1997-07-16T19:20:30.45+01:00',
-            869077230,
-        ];
-
-        yield 'W3CDTF Example 1' => [
-            '1994-11-05T08:15:30-05:00',
-            784041330,
-        ];
-        yield 'W3CDTF Example 2' => [
-            '1994-11-05T13:15:30Z',
-            784041330,
-        ];
-
-        yield 'bug 259 test 0' => [
-            '1994-11-05T08:15:30-0500',
-            784041330,
-        ];
-    }
-
     /**
-     * @dataProvider w3cDtfDatesProvider
+     * Set the Registry into the class
+     *
+     * @param Registry $registry
+     *
+     * @return void
      */
-    public function testW3cDtf($data, $expected)
-    {
-        $date = new Date();
-        $this->assertSame(
-            $expected,
-            $date->date_w3cdtf($data)
-        );
-    }
+    public function set_registry(Registry $registry)/* : void */;
 }