+++ src/Source.php
@@ -53,7 +53,7 @@ namespace SimplePie;
  * @package SimplePie
  * @subpackage API
  */
-class Source
+class Source implements RegistryAware
 {
     public $item;
     public $data = [];
@@ -65,7 +65,7 @@ class Source
         $this->data = $data;
     }
 
-    public function set_registry(\SimplePie\Registry $registry)
+    public function set_registry(\SimplePie\Registry $registry)/* : void */
     {
         $this->registry = $registry;
     }